Switch to new Gradle bundlor plugin
Remove all template.mf files in favor of concise inlining in the build
script (see diff). Build now fails if Bundlor raises any warnings, i.e.
that a given import has not been declared. Note that many such warnings
were raised when performing this refactoring, and that they have each
been fixed.

This change works, but requires that the BundlorPlugin.groovy and
bundlor.properties files under buildSrc be removed. This is because
the buildSrc version of the plugin conflicts with the jar-based
version that is downloaded from a remote repository. This problem
will go away once buildSrc is removed entirely.